+++ src/lib.rs
@@ -17,10 +17,15 @@
 #![allow(clippy::cargo_common_metadata)]
 #![allow(clippy::separated_literal_suffix)]
 #![allow(clippy::float_arithmetic)]
+#![allow(clippy::struct_excessive_bools)]
 
 use anyhow::{anyhow, Error, Result};
 use clap::Parser;
+use dialoguer::Input;
 use dirs::home_dir;
+use log::info;
+use prettytable::{format, Cell, Row, Table};
+use serde_json::json;
 use sled::Db;
 use std::fmt::{Display, Formatter};
 use std::str::FromStr;
@@ -28,8 +33,8 @@ use std::str::FromStr;
 /// The prelude for the `ask-bayes` crate.
 pub mod prelude {
     pub use crate::{
-        calculate_posterior_probability, get_prior, remove_prior, set_prior, Args, Evidence,
-        UpdateHypothesis,
+        calculate_posterior_probability, get_prior, remove_prior, report_posterior_probability,
+        set_prior, wizard, Args, Evidence, UpdateHypothesis,
     };
 }
 
@@ -49,8 +54,8 @@ impl FromStr for Evidence {
     #[inline]
     fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
         match s {
-            "o" | "observed" | "Observed" => Ok(Self::Observed),
-            "n" | "not-observed" | "NotObserved" => Ok(Self::NotObserved),
+            "o" | "observed" | "Observed" | "y" | "Y" => Ok(Self::Observed),
+            "n" | "not-observed" | "NotObserved" | "N" | "not observed" => Ok(Self::NotObserved),
             _ => Err(anyhow!("Invalid evidence: {}", s)),
         }
     }
@@ -82,8 +87,8 @@ impl FromStr for UpdateHypothesis {
     #[inline]
     fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
         match s {
-            "u" | "update" | "Update" => Ok(Self::Update),
-            "n" | "no-update" | "NoUpdate" => Ok(Self::NoUpdate),
+            "u" | "update" | "Update" | "y" | "Y" => Ok(Self::Update),
+            "n" | "no-update" | "NoUpdate" | "N" => Ok(Self::NoUpdate),
             _ => Err(anyhow!("Invalid update hypothesis: {}", s)),
         }
     }
@@ -99,29 +104,100 @@ impl Display for UpdateHypothesis {
     }
 }
 
+/// The output format to use when displaying results to the user
+#[derive(Debug, Clone, PartialEq, Eq)]
+#[non_exhaustive]
+pub enum OutputFormat {
+    /// Output in a table format
+    Table,
+    /// Output in a JSON format
+    Json,
+    /// Output in a formatted string
+    Simple,
+}
+
+impl FromStr for OutputFormat {
+    type Err = Error;
+
+    #[inline]
+    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
+        match s {
+            "table" | "Table" | "t" | "T" => Ok(Self::Table),
+            "json" | "Json" | "j" | "J" => Ok(Self::Json),
+            "simple" | "Simple" | "s" | "S" => Ok(Self::Simple),
+            _ => Err(anyhow!("Invalid output format: {}", s)),
+        }
+    }
+}
+
+impl Display for OutputFormat {
+    #[inline]
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match *self {
+            Self::Table => write!(f, "Table"),
+            Self::Json => write!(f, "Json"),
+            Self::Simple => write!(f, "Simple"),
+        }
+    }
+}
+
 /// Arguments for the `ask-bayes` command
 #[derive(Parser, Debug)]
 #[non_exhaustive]
 #[clap(author, version, about, long_about = None)]
 pub struct Args {
     /// Name of the Hypothesis to update
-    #[clap(short, long)]
-    pub name: String,
+    #[clap(
+        short,
+        long,
+        forbid_empty_values = true,
+        required_unless_present("wizard")
+    )]
+    pub name: Option<String>,
     /// The prior probability of the hypothesis P(H)
-    #[clap(short, long, default_value_t = 0.5, validator = validate_probability)]
-    pub prior: f64,
+    #[clap(
+        short,
+        long,
+        default_value_if("name", None, Some("0.5")),
+        validator = parse_validate_probability,
+        forbid_empty_values = true,
+        required_unless_present("wizard")
+    )]
+    pub prior: Option<f64>,
     /// The likelihood of the evidence P(E|H)
-    #[clap(short, long, default_value_t = 0.5, validator = validate_probability)]
-    pub likelihood: f64,
+    #[clap(
+        short,
+        long,
+        default_value_if("name", None, Some("0.5")),
+        validator = parse_validate_probability,
+        forbid_empty_values = true,
+        required_unless_present("wizard"))]
+    pub likelihood: Option<f64>,
     /// The likelihood of the evidence P(E|¬H)
-    #[clap(long, default_value_t = 0.5, validator = validate_probability)]
-    pub likelihood_not: f64,
+    #[clap(
+        long,
+        default_value_if("name", None, Some("0.5")),
+        validator = parse_validate_probability,
+        forbid_empty_values = true,
+        required_unless_present("wizard"))]
+    pub likelihood_null: Option<f64>,
     /// Indicates whether supporting evidence is observed
-    #[clap(short, long, default_value_t = Evidence::Observed)]
-    pub evidence: Evidence,
+    #[clap(
+        short,
+        long,
+        default_value_if("name", None, Some("Observed")),
+        default_missing_value = "Observed",
+        possible_values = ["o", "observed", "Observed", "n", "not-observed", "NotObserved"],
+        required_unless_present("wizard"))]
+    pub evidence: Option<Evidence>,
     /// Updates the prior probability of the hypothesis P(H) to the new posterior probability, saving it to the database
-    #[clap(short, long, default_value_t = UpdateHypothesis::NoUpdate)]
-    pub update_prior: UpdateHypothesis,
+    #[clap(
+        short,
+        long,
+        default_value_if("name", None, Some("NoUpdate")),
+        default_missing_value = "Update",
+        possible_values = ["u", "update", "Update", "n", "no-update", "NoUpdate"])]
+    pub update_prior: Option<UpdateHypothesis>,
     /// Returns the saved value of the prior probability of the hypothesis P(H).
     /// Incompatible with other flags aside from `--name`
     #[clap(
@@ -129,7 +205,7 @@ pub struct Args {
         long,
         conflicts_with = "prior",
         conflicts_with = "likelihood",
-        conflicts_with = "likelihood-not",
+        conflicts_with = "likelihood-null",
         conflicts_with = "evidence",
         conflicts_with = "update-prior"
     )]
@@ -139,14 +215,14 @@ pub struct Args {
     #[clap(
         short,
         long,
-        requires = "prior",
+        conflicts_with = "prior",
         conflicts_with = "likelihood",
-        conflicts_with = "likelihood-not",
+        conflicts_with = "likelihood-null",
         conflicts_with = "evidence",
         conflicts_with = "update-prior",
         conflicts_with = "get-prior"
     )]
-    pub set_prior: bool,
+    pub set_prior: Option<f64>,
     /// Removes the prior probability of the hypothesis P(H) from the database.
     /// Incompatible with other flags aside from `--name`
     #[clap(
@@ -154,13 +230,25 @@ pub struct Args {
         long,
         conflicts_with = "prior",
         conflicts_with = "likelihood",
-        conflicts_with = "likelihood-not",
+        conflicts_with = "likelihood-null",
         conflicts_with = "evidence",
         conflicts_with = "update-prior",
         conflicts_with = "set-prior",
         conflicts_with = "get-prior"
     )]
     pub remove_prior: bool,
+    /// Runs the wizard to help guide you through the process of updating a hypothesis
+    #[clap(short, long, exclusive = true, takes_value = false)]
+    pub wizard: bool,
+    /// The type of output to display
+    #[clap(
+        short,
+        long,
+        default_value_if("name", None, Some("Table")),
+        possible_values = ["t", "table", "Table", "T", "j", "json", "Json", "J", "s", "simple", "Simple", "S"],
+        required_unless_present("wizard")
+    )]
+    pub output: Option<OutputFormat>,
 }
 
 /// The posterior probability of the hypothesis P(H|E) if the evidence is observed, or P(H|¬E) if the evidence is not observed
@@ -170,12 +258,12 @@ pub struct Args {
 pub fn calculate_posterior_probability(
     prior: f64,
     likelihood: f64,
-    likelihood_not: f64,
+    likelihood_null: f64,
     evidence: &Evidence,
     name: &str,
 ) -> Result<f64> {
-    validate_likelihoods_and_prior(prior, likelihood, likelihood_not, evidence, name)?;
-    let p_e = probability_of_observing_evidence(prior, likelihood, likelihood_not);
+    validate_likelihoods_and_prior(prior, likelihood, likelihood_null, evidence, name)?;
+    let p_e = marginal_likelihood(prior, likelihood, likelihood_null);
     match *evidence {
         Evidence::Observed => {
             // P(H|E) = P(H) * P(E|H) / P(E)
@@ -245,13 +333,19 @@ fn open_db() -> Result<Db> {
     Ok(sled::open(db_path)?)
 }
 
-/// Validates a probability.  Probabilities should be valid floats between 0 and 1.
-fn validate_probability(value: &str) -> Result<f64> {
+/// Parses and validates a probability
+fn parse_validate_probability(value: &str) -> Result<f64> {
     let float = value.parse::<f64>()?;
-    if !(0.0_f64..=1.0_f64).contains(&float) {
+    validate_probability(float)?;
+    Ok(float)
+}
+
+/// Validates a probability.  Probabilities should be valid floats between 0 and 1.
+fn validate_probability(value: f64) -> Result<()> {
+    if !(0.0_f64..=1.0_f64).contains(&value) {
         return Err(anyhow!("Probability must be between 0 and 1"));
     }
-    Ok(float)
+    Ok(())
 }
 
 /// Negates a probability.  Ex. P(H) -> P(¬H)
@@ -263,24 +357,19 @@ fn negate(value: f64) -> f64 {
 fn validate_likelihoods_and_prior(
     prior: f64,
     likelihood: f64,
-    likelihood_not: f64,
+    likelihood_null: f64,
     evidence: &Evidence,
     name: &str,
 ) -> Result<()> {
     match *evidence {
         Evidence::Observed => {
-            if probability_of_observing_evidence(prior, likelihood, likelihood_not) <= 0.0_f64 {
-                return Err(anyhow!("The total probability of observing evidence P(E) must be greater than 0 if evidence is observed.  \r\nP(E) = P({name})[{prior}] * P(E|{name})[{likelihood}] + P(\u{ac}{name})[{}] * P(E|\u{ac}{name})[{}] = 0", negate(prior), likelihood_not));
+            if marginal_likelihood(prior, likelihood, likelihood_null) <= 0.0_f64 {
+                return Err(anyhow!("The total probability of observing evidence P(E) must be greater than 0 if evidence is observed.  \r\nP(E) = P({name})[{prior}] * P(E|{name})[{likelihood}] + P(\u{ac}{name})[{}] * P(E|\u{ac}{name})[{}] = 0", negate(prior), likelihood_null));
             }
         }
         Evidence::NotObserved => {
-            if negate(probability_of_observing_evidence(
-                prior,
-                likelihood,
-                likelihood_not,
-            )) <= 0.0_f64
-            {
-                return Err(anyhow!("The total probability of not observing evidence P(\u{ac}E) must be greater than 0 if evidence is not observed.  \r\nP(\u{ac}E) = P(\u{ac}E|{name})[{}] * P({name})[{prior}] + P(\u{ac}{name})[{}] * P(\u{ac}E|\u{ac}{name})[{}] = 0", negate(likelihood), negate(prior), negate(likelihood_not)));
+            if negate(marginal_likelihood(prior, likelihood, likelihood_null)) <= 0.0_f64 {
+                return Err(anyhow!("The total probability of not observing evidence P(\u{ac}E) must be greater than 0 if evidence is not observed.  \r\nP(\u{ac}E) = P(\u{ac}E|{name})[{}] * P({name})[{prior}] + P(\u{ac}{name})[{}] * P(\u{ac}E|\u{ac}{name})[{}] = 0", negate(likelihood), negate(prior), negate(likelihood_null)));
             }
         }
     }
@@ -288,95 +377,296 @@ fn validate_likelihoods_and_prior(
     Ok(())
 }
 
-///  P(H) * P(E|H) + P(¬H) * P(E|¬H)
-fn probability_of_observing_evidence(prior: f64, likelihood: f64, likelihood_not: f64) -> f64 {
-    likelihood.mul_add(prior, likelihood_not * negate(prior))
+///  P(H) * P(E|H) + P(¬H) * P(E|¬H), otherwise known as P(E)
+fn marginal_likelihood(prior: f64, likelihood: f64, likelihood_null: f64) -> f64 {
+    likelihood.mul_add(prior, likelihood_null * negate(prior))
+}
+
+/// Runs the wizard to guide the update of the prior probability of the hypothesis
+/// # Errors
+/// - If the prompt cannot be displayed
+#[inline]
+#[cfg(not(tarpaulin_include))]
+pub fn wizard() -> Result<()> {
+    let name = Input::<String>::new()
+        .with_prompt("Enter the name of the hypothesis")
+        .allow_empty(false)
+        .interact_text()?;
+
+    let prior = Input::<f64>::new()
+        .with_prompt(format!(
+            "Enter the prior probability of the hypothesis P({name})"
+        ))
+        .allow_empty(false)
+        .default(0.5_f64)
+        .validate_with(|v: &f64| validate_probability(*v))
+        .interact_text()?;
+
+    let likelihood = Input::<f64>::new()
+        .with_prompt(format!(
+            "Enter the likelihood of observing evidence given {name} is true P(E|{name})"
+        ))
+        .allow_empty(false)
+        .default(0.5_f64)
+        .validate_with(|v: &f64| validate_probability(*v))
+        .interact_text()?;
+
+    let likelihood_null = Input::<f64>::new()
+        .with_prompt(format!(
+            "Enter the likelihood of observing evidence given {name} is false P(E|\u{ac}{name})"
+        ))
+        .allow_empty(false)
+        .default(0.5_f64)
+        .validate_with(|v: &f64| validate_probability(*v))
+        .interact_text()?;
+
+    let evidence = Input::<Evidence>::new()
+        .with_prompt("Is evidence observed or not observed?".to_owned())
+        .allow_empty(false)
+        .default(Evidence::Observed)
+        .interact_text()?;
+
+    let posterior_probability =
+        calculate_posterior_probability(prior, likelihood, likelihood_null, &evidence, &name)?;
+
+    let output_format = Input::<OutputFormat>::new()
+        .with_prompt("How would you like the output?".to_owned())
+        .allow_empty(false)
+        .default(OutputFormat::Table)
+        .interact_text()?;
+
+    report_posterior_probability(
+        prior,
+        likelihood,
+        likelihood_null,
+        &evidence,
+        posterior_probability,
+        &name,
+        &output_format,
+    );
+
+    let update = Input::<UpdateHypothesis>::new()
+        .with_prompt("Would you like to update the prior probability?".to_owned())
+        .allow_empty(false)
+        .default(UpdateHypothesis::NoUpdate)
+        .interact_text()?;
+
+    if update == UpdateHypothesis::Update {
+        set_prior(&name, posterior_probability)?;
+        info!("P({name}) has been updated to {}", posterior_probability);
+    }
+
+    Ok(())
+}
+
+/// Reports the posterior probability of the hypothesis given the evidence.  Also reports the values of the `prior`, `likelihood`, and `likelihood_null`.
+#[inline]
+#[cfg(not(tarpaulin_include))]
+pub fn report_posterior_probability(
+    prior: f64,
+    likelihood: f64,
+    likelihood_null: f64,
+    evidence: &Evidence,
+    posterior_probability: f64,
+    name: &str,
+    output_format: &OutputFormat,
+) {
+    match *output_format {
+        OutputFormat::Table => {
+            report_table(
+                name,
+                prior,
+                likelihood,
+                likelihood_null,
+                evidence,
+                posterior_probability,
+            );
+        }
+        OutputFormat::Json => {
+            report_json(
+                name,
+                prior,
+                likelihood,
+                likelihood_null,
+                evidence,
+                posterior_probability,
+            );
+        }
+        OutputFormat::Simple => {
+            let output = format!(
+                "
+                P({name}) = {prior}
+                P(E|{name}) = {likelihood}
+                P(E|\u{ac}{name}) = {likelihood_null}
+                P({name}|E) = {posterior_probability}
+                "
+            );
+            info!("{output}");
+        }
+    }
+}
+
+/// Reports the posterior probability of the hypothesis given the evidence in a table format.
+#[cfg(not(tarpaulin_include))]
+fn report_table(
+    name: &str,
+    prior: f64,
+    likelihood: f64,
+    likelihood_null: f64,
+    evidence: &Evidence,
+    posterior_probability: f64,
+) {
+    let marginal_likelihood = marginal_likelihood(prior, likelihood, likelihood_null);
+    let mut table = Table::new();
+    table.set_format(*format::consts::FORMAT_NO_LINESEP_WITH_TITLE);
+    table.set_titles(Row::new(vec![
+        Cell::new("Name"),
+        Cell::new("Probability"),
+        Cell::new("Value"),
+    ]));
+    table.add_row(Row::new(vec![
+        Cell::new("Prior"),
+        Cell::new(&format!("P({name})")),
+        Cell::new(&format!("{prior}")),
+    ]));
+    table.add_row(Row::new(vec![
+        Cell::new("Likelihood"),
+        Cell::new(&format!("P(E|{name})")),
+        Cell::new(&format!("{likelihood}")),
+    ]));
+    table.add_row(Row::new(vec![
+        Cell::new("Likelihood Null"),
+        Cell::new(&format!("P(E|\u{ac}{name})")),
+        Cell::new(&format!("{likelihood_null}")),
+    ]));
+    table.add_row(Row::new(vec![
+        Cell::new("Marginal Likelihood"),
+        Cell::new("P(E)"),
+        Cell::new(&format!("{marginal_likelihood}")),
+    ]));
+
+    match *evidence {
+        Evidence::Observed => table.add_row(Row::new(vec![
+            Cell::new("Posterior Probability"),
+            Cell::new(&format!("P({name}|E)")),
+            Cell::new(&format!("{posterior_probability}")),
+        ])),
+        Evidence::NotObserved => table.add_row(Row::new(vec![
+            Cell::new(&format!("P({name}|\u{ac}E)")),
+            Cell::new(&format!("{posterior_probability}")),
+        ])),
+    };
+
+    table.printstd();
+}
+
+/// Reports the posterior probability of the hypothesis given the evidence in a JSON format.
+#[cfg(not(tarpaulin_include))]
+fn report_json(
+    name: &str,
+    prior: f64,
+    likelihood: f64,
+    likelihood_null: f64,
+    evidence: &Evidence,
+    posterior_probability: f64,
+) {
+    let json = json!({
+        "name": name,
+        "prior": prior,
+        "likelihood": likelihood,
+        "likelihood_null": likelihood_null,
+        "evidence": match *evidence {
+            Evidence::Observed => "observed",
+            Evidence::NotObserved => "not observed",
+        },
+        "posterior_probability": posterior_probability,
+    });
+
+    info!("{}", json.to_string());
 }
 
 #[cfg(test)]
+#[allow(clippy::panic_in_result_fn)]
 mod tests {
     use super::*;
+
+    fn epsilon_compare(a: f64, b: f64) -> bool {
+        (a - b).abs() < f64::EPSILON
+    }
 
     #[test]
     fn it_validates_a_valid_probability() -> Result<()> {
         let prob = "0.75";
-        let result = validate_probability(prob)?;
-        assert_eq!(result, 0.75_f64);
+        let result = parse_validate_probability(prob)?;
+        assert!(epsilon_compare(result, 0.75_f64));
         Ok(())
     }
 
     #[test]
-    fn it_fails_to_validate_a_probability_greater_than_1() -> Result<()> {
+    fn it_fails_to_validate_a_probability_greater_than_1() {
         let prob = "1.1";
-        let result = validate_probability(prob);
+        let result = parse_validate_probability(prob);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
-    fn it_fails_to_validate_a_probability_less_than_0() -> Result<()> {
+    fn it_fails_to_validate_a_probability_less_than_0() {
         let prob = "-0.1";
-        let result = validate_probability(prob);
+        let result = parse_validate_probability(prob);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
-    fn it_fails_to_validate_an_invalid_float() -> Result<()> {
+    fn it_fails_to_validate_an_invalid_float() {
         let prob = "invalid";
-        let result = validate_probability(prob);
+        let result = parse_validate_probability(prob);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
     fn it_validates_a_valid_pair_of_likelihoods() -> Result<()> {
-        let likelihood = 0.75;
-        let likelihood_not = 0.25;
-        let prior = 0.5;
+        let likelihood = 0.75_f64;
+        let likelihood_null = 0.25_f64;
+        let prior = 0.5_f64;
         let evidence = Evidence::Observed;
         let name = "test";
-        validate_likelihoods_and_prior(prior, likelihood, likelihood_not, &evidence, name)?;
-        Ok(())
+        validate_likelihoods_and_prior(prior, likelihood, likelihood_null, &evidence, name)
     }
 
     #[test]
     fn it_validates_a_valid_pair_of_negated_likelihoods() -> Result<()> {
-        let prior = 0.5;
-        let likelihood = 0.75;
-        let likelihood_not = 0.25;
+        let prior = 0.5_f64;
+        let likelihood = 0.75_f64;
+        let likelihood_null = 0.25_f64;
         let evidence = Evidence::NotObserved;
         let name = "test";
-        validate_likelihoods_and_prior(prior, likelihood, likelihood_not, &evidence, name)?;
-        Ok(())
+        validate_likelihoods_and_prior(prior, likelihood, likelihood_null, &evidence, name)
     }
 
     #[test]
     fn it_fails_to_validate_a_pair_of_likelihoods_with_evidence_observed_when_the_sum_is_less_than_or_equal_to_0(
-    ) -> Result<()> {
-        let prior = 0.5;
-        let likelihood = 0.0;
-        let likelihood_not = 0.0;
+    ) {
+        let prior = 0.5_f64;
+        let likelihood = 0.0_f64;
+        let likelihood_null = 0.0_f64;
         let evidence = Evidence::Observed;
         let name = "test";
         let result =
-            validate_likelihoods_and_prior(prior, likelihood, likelihood_not, &evidence, name);
+            validate_likelihoods_and_prior(prior, likelihood, likelihood_null, &evidence, name);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
     fn it_fails_to_validate_a_pair_of_negated_likelihoods_with_evidence_not_observed_when_the_negated_sum_is_less_than_or_equal_to_0(
-    ) -> Result<()> {
-        let prior = 0.5;
-        let likelihood = 1.0;
-        let likelihood_not = 1.0;
+    ) {
+        let prior = 0.5_f64;
+        let likelihood = 1.0_f64;
+        let likelihood_null = 1.0_f64;
         let evidence = Evidence::NotObserved;
         let name = "test";
         let result =
-            validate_likelihoods_and_prior(prior, likelihood, likelihood_not, &evidence, name);
+            validate_likelihoods_and_prior(prior, likelihood, likelihood_null, &evidence, name);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
@@ -415,15 +705,14 @@ mod tests {
     }
 
     #[test]
-    fn it_fails_to_parse_an_invalid_evidence_string() -> Result<()> {
+    fn it_fails_to_parse_an_invalid_evidence_string() {
         let evidence = "invalid";
         let result = Evidence::from_str(evidence);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
-    fn it_displays_a_valid_evidence_string() -> Result<()> {
+    fn it_displays_a_valid_evidence_string() {
         {
             let evidence = Evidence::Observed;
             let result = evidence.to_string();
@@ -434,7 +723,6 @@ mod tests {
             let result = evidence.to_string();
             assert_eq!(result, "NotObserved");
         }
-        Ok(())
     }
 
     #[test]
@@ -473,15 +761,14 @@ mod tests {
     }
 
     #[test]
-    fn it_fails_to_parse_an_invalid_update_string() -> Result<()> {
+    fn it_fails_to_parse_an_invalid_update_string() {
         let update = "invalid";
         let result = UpdateHypothesis::from_str(update);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
-    fn it_displays_a_valid_update_string() -> Result<()> {
+    fn it_displays_a_valid_update_string() {
         {
             let update = UpdateHypothesis::Update;
             let result = update.to_string();
@@ -492,59 +779,152 @@ mod tests {
             let result = update.to_string();
             assert_eq!(result, "NoUpdate");
         }
-        Ok(())
     }
 
     #[test]
     fn it_calculates_the_posterior_probability_when_evidence_is_observed() -> Result<()> {
-        let prior = 0.75;
-        let likelihood = 0.75;
-        let likelihood_not = 0.5;
+        let prior = 0.75_f64;
+        let likelihood = 0.75_f64;
+        let likelihood_null = 0.5_f64;
         let evidence = Evidence::Observed;
         let name = "test";
         let result =
-            calculate_posterior_probability(prior, likelihood, likelihood_not, &evidence, name)?;
-        assert_eq!(result, 0.8181818181818182);
+            calculate_posterior_probability(prior, likelihood, likelihood_null, &evidence, name)?;
+        assert!(epsilon_compare(result, 0.818_181_818_181_818_2_f64));
         Ok(())
     }
 
     #[test]
     fn it_calculates_the_posterior_probability_when_evidence_is_not_observed() -> Result<()> {
-        let prior = 0.75;
-        let likelihood = 0.75;
-        let likelihood_not = 0.5;
+        let prior = 0.75_f64;
+        let likelihood = 0.75_f64;
+        let likelihood_null = 0.5_f64;
         let evidence = Evidence::NotObserved;
         let name = "test";
         let result =
-            calculate_posterior_probability(prior, likelihood, likelihood_not, &evidence, name)?;
-        assert_eq!(result, 0.6);
+            calculate_posterior_probability(prior, likelihood, likelihood_null, &evidence, name)?;
+        assert!(epsilon_compare(result, 0.6));
         Ok(())
     }
 
     #[test]
-    fn it_fails_to_validate_likelihoods_and_hypothesis_when_the_negated_prior_is_zero() -> Result<()>
-    {
+    fn it_fails_to_validate_likelihoods_and_hypothesis_when_the_negated_prior_is_zero() {
         let name = "test";
-        let prior = 1.0;
-        let likelihood = 1.0;
-        let likelihood_not = 0.5;
+        let prior = 1.0_f64;
+        let likelihood = 1.0_f64;
+        let likelihood_null = 0.5_f64;
         let evidence = Evidence::NotObserved;
         let result =
-            validate_likelihoods_and_prior(prior, likelihood, likelihood_not, &evidence, name);
+            validate_likelihoods_and_prior(prior, likelihood, likelihood_null, &evidence, name);
         assert!(result.is_err());
-        Ok(())
     }
 
     #[test]
-    fn it_fails_to_validate_likelihoods_and_hypothesis_when_the_prior_is_zero() -> Result<()> {
+    fn it_fails_to_validate_likelihoods_and_hypothesis_when_the_prior_is_zero() {
         let name = "test";
-        let prior = 0.0;
-        let likelihood = 0.5;
-        let likelihood_not = 0.0;
+        let prior = 0.0_f64;
+        let likelihood = 0.5_f64;
+        let likelihood_null = 0.0_f64;
         let evidence = Evidence::Observed;
         let result =
-            validate_likelihoods_and_prior(prior, likelihood, likelihood_not, &evidence, name);
+            validate_likelihoods_and_prior(prior, likelihood, likelihood_null, &evidence, name);
         assert!(result.is_err());
+    }
+
+    #[test]
+    fn it_parses_a_valid_output_format() -> Result<()> {
+        {
+            let format = "json";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Json);
+        }
+        {
+            let format = "j";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Json);
+        }
+        {
+            let format = "Json";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Json);
+        }
+        {
+            let format = "J";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Json);
+        }
+        {
+            let format = "t";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Table);
+        }
+        {
+            let format = "Table";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Table);
+        }
+        {
+            let format = "table";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Table);
+        }
+        {
+            let format = "T";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Table);
+        }
+        {
+            let format = "simple";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Simple);
+        }
+        {
+            let format = "s";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Simple);
+        }
+        {
+            let format = "S";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Simple);
+        }
+        {
+            let format = "simple";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Simple);
+        }
+        {
+            let format = "Simple";
+            let result = OutputFormat::from_str(format)?;
+            assert_eq!(result, OutputFormat::Simple);
+        }
+
         Ok(())
+    }
+
+    #[test]
+    fn it_fails_to_parse_an_invalid_output_format() {
+        let format = "invalid";
+        let result = OutputFormat::from_str(format);
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn it_displays_a_valid_output_format() {
+        {
+            let format = OutputFormat::Json;
+            let result = format.to_string();
+            assert_eq!(result, "Json");
+        }
+        {
+            let format = OutputFormat::Table;
+            let result = format.to_string();
+            assert_eq!(result, "Table");
+        }
+        {
+            let format = OutputFormat::Simple;
+            let result = format.to_string();
+            assert_eq!(result, "Simple");
+        }
     }
 }